requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
